+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Ensure you use consistent title format.
 
 - ui: Remove support for 7.x agents.
 - ui: Accept editing an offline instance.
+- ui: Search for .config/temboard.conf.
 - Remove python 2.7 support.
 - Remove daemonization. Use nohup or systemd.
 - Dropped pluging hotplug.